learned?
To avoid the unhashable dict bug, I make the "state" string before making a dictionary of dictionary. In other word, in the previous version, I use the state(tuple or dict) as the key to the value (another dict). Now I use state(string) as the key to the value (another dict.
There are two points should be noted: 1. I think the previous version the state should be 'tuple', but it turns out to be dict.
2. to avoid some coding problem, I do the stringization in every def of the class.
PS. I still cant understand why the minor violation rate is soaring.